dep-updates: Move getPdfJs
